+++ src/models/gtfs/frequencies.ts
@@ -11,22 +11,14 @@ export const frequencies = {
     },
     {
       name: 'start_time',
-      type: 'text',
+      type: 'time',
       required: true,
       primary: true,
     },
     {
-      name: 'start_timestamp',
-      type: 'integer',
-    },
-    {
       name: 'end_time',
-      type: 'text',
+      type: 'time',
       required: true,
-    },
-    {
-      name: 'end_timestamp',
-      type: 'integer',
     },
     {
       name: 'headway_secs',