+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Updated
-- Renamed time caching functions
+- Updates to time columns and timestamp columns
 
 ### Fixed
 - Better GTFS export for currency